refactor: 유효성 검증 로직 과 에러 상태 분리
