+++ src/types/models/FormErrors.ts
@@ -1,0 +1,11 @@
+export interface FormErrors {
+  image: string;
+  name: string;
+  phoneNumber: string;
+  birthDate: string;
+  gender: string;
+  availableStartDate: string;
+  selfIntroduction: string;
+  mbti: string;
+  instagramId: string;
+}
